pyproject.toml
[tool.poetry]
name = "nannernest"
version = "0.1.0"
description = "Arrange banana slices on a piece of bread in order to optimize banana coverage."
authors = ["EthanRosenthal <ethan@ethanrosenthal.com>"]

[tool.poetry.scripts]
nannernest = 'nannernest.cli:app'

[tool.poetry.dependencies]
python = ">=3.6.1"
typer = "^0.3.1"
sympy = "^1.6.1"
seaborn = "^0.10.1"
numpy = "^1.19.1"
matplotlib = "^3.3.0"
svgpath2mpl = "^0.2.1"
Pillow = "^7.2.0"
torch = "^1.6.0"
torchvision = "^0.7.0"
scipy = "^1.5.2"
scikit-image = "^0.17.2"
nest2D = { git = "git@github.com:EthanRosenthal/nest2D.git", branch = "download-dependencies" }
colorama = "^0.4.3"
click-spinner = "^0.1.10"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
pytest = "^6.0.1"
ipykernel = "^5.3.4"
pytest-cov = "^2.10.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
